% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalibreMD.R
\name{explode_tags}
\alias{explode_tags}
\title{Explode tags}
\usage{
explode_tags(eav)
}
\arguments{
\item{eav}{The EAV data frame}
}
\value{
The EAV data frame with exploded tags
}
\description{
Explode tags
}
