% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalibreMD.R
\name{explode_text_features}
\alias{explode_text_features}
\title{Explode text features into individual tokens}
\usage{
explode_text_features(eav)
}
\arguments{
\item{eav}{The EAV data frame}
}
\value{
A data frame with text features exploded into tokens
}
\description{
Explode text features into individual tokens
}
