% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalibreMD.R
\name{find_md_db}
\alias{find_md_db}
\title{Find the metadata database file in the given directory}
\usage{
find_md_db(dataDir)
}
\arguments{
\item{dataDir}{The directory containing the Calibre library}
}
\value{
The path to the metadata.db file
}
\description{
Find the metadata database file in the given directory
}
