% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalibreMD.R
\name{get_book_titles}
\alias{get_book_titles}
\title{Get the titles of all books}
\usage{
get_book_titles(eav)
}
\arguments{
\item{eav}{The EAV data frame}
}
\value{
A data frame with book IDs and titles
}
\description{
Get the titles of all books
}
