% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalibreMD.R
\name{get_existing_tags}
\alias{get_existing_tags}
\title{Get the existing tags for each book}
\usage{
get_existing_tags(eav)
}
\arguments{
\item{eav}{The EAV data frame}
}
\value{
A data frame with book IDs and their existing tags
}
\description{
Get the existing tags for each book
}
