% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalibreMD.R
\name{get_label_probabilities}
\alias{get_label_probabilities}
\title{Get label probabilities in long format}
\usage{
get_label_probabilities(predictions_probs)
}
\arguments{
\item{predictions_probs}{Prediction probabilities}
}
\value{
A data frame of probabilities in long format
}
\description{
Get label probabilities in long format
}
