% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalibreMD.R
\name{get_recommended_add}
\alias{get_recommended_add}
\title{Get recommended tags to add}
\usage{
get_recommended_add(eav, pred_long_all, add_threshold)
}
\arguments{
\item{eav}{EAV data frame}

\item{pred_long_all}{Predictions in long format}

\item{add_threshold}{Threshold for adding tags}
}
\value{
A data frame of recommended tags to add
}
\description{
Get recommended tags to add
}
