% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalibreMD.R
\name{get_recommended_remove}
\alias{get_recommended_remove}
\title{Get books to remove a specific tag from}
\usage{
get_recommended_remove(eav, pred_long_all, remove_threshold = 0.05)
}
\arguments{
\item{eav}{EAV data frame}

\item{pred_long_all}{Predictions in long format}

\item{remove_threshold}{Threshold for removing tags}
}
\value{
A data frame of recommended tags to remove
}
\description{
Get books to remove a specific tag from
}
