% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalibreMD.R
\name{get_tag_counts}
\alias{get_tag_counts}
\title{Get the count of tags per book}
\usage{
get_tag_counts(eav)
}
\arguments{
\item{eav}{The EAV data frame}
}
\value{
A data frame with book IDs and their tag counts
}
\description{
Get the count of tags per book
}
