% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalibreMD.R
\name{get_tag_summary}
\alias{get_tag_summary}
\title{Get a summary of tags and their book counts}
\usage{
get_tag_summary(eav)
}
\arguments{
\item{eav}{The EAV data frame}
}
\value{
A data frame with tags and their book counts
}
\description{
Get a summary of tags and their book counts
}
