% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalibreMD.R
\name{get_top_n_labels}
\alias{get_top_n_labels}
\title{Get top N labels for a book}
\usage{
get_top_n_labels(pred_probs, book_id, n = 2)
}
\arguments{
\item{pred_probs}{Prediction probabilities}

\item{book_id}{Book ID to get labels for}

\item{n}{Number of top labels to return}
}
\value{
Character vector of top N labels
}
\description{
Get top N labels for a book
}
