% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalibreMD.R
\name{load_eav}
\alias{load_eav}
\title{Load the EAV (Entity-Attribute-Value) data from the Calibre metadata database}
\usage{
load_eav(dbname)
}
\arguments{
\item{dbname}{The path to the metadata.db file}
}
\value{
A data frame containing the EAV data
}
\description{
Load the EAV (Entity-Attribute-Value) data from the Calibre metadata database
}
