% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalibreMD.R
\name{prep_dataset}
\alias{prep_dataset}
\title{Prepare the dataset for modeling}
\usage{
prep_dataset(eav)
}
\arguments{
\item{eav}{The EAV data frame}
}
\value{
A list containing the feature matrix, tag matrix, and book features
}
\description{
Prepare the dataset for modeling
}
