% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalibreMD.R
\name{suggest_tag_additions}
\alias{suggest_tag_additions}
\title{Suggest tag additions for a specific tag}
\usage{
suggest_tag_additions(eav, predictions = pred_long_all, tag_name, prob_min = 0)
}
\arguments{
\item{eav}{EAV data frame}

\item{predictions}{Predictions in long format}

\item{tag_name}{Tag to suggest additions for}

\item{prob_min}{Minimum probability threshold}
}
\value{
A data frame of suggested tag additions
}
\description{
Suggest tag additions for a specific tag
}
