% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalibreMD.R
\name{threshold_predictions}
\alias{threshold_predictions}
\title{Threshold predictions to get discrete labels}
\usage{
threshold_predictions(pred_probs, book_id, threshold = 0.5)
}
\arguments{
\item{pred_probs}{Prediction probabilities}

\item{book_id}{Book ID to threshold}

\item{threshold}{Probability threshold}
}
\value{
Character vector of predicted tags
}
\description{
Threshold predictions to get discrete labels
}
