% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalibreMD.R
\name{train_models}
\alias{train_models}
\title{Train models for tag prediction}
\usage{
train_models(dataset)
}
\arguments{
\item{dataset}{A list containing feature matrix and tag matrix}
}
\value{
A list of trained models
}
\description{
Train models for tag prediction
}
